prototype/requirements.txt
# Weaver Forge Python Dependencies

# CLI framework
typer>=0.9.0

# Data validation and LLM integration  
pydantic>=2.5.0
pydantic-ai>=0.0.8

# Async file operations
aiofiles>=23.2.1

# YAML processing
pyyaml>=6.0.1

# Contract-based programming
icontract>=2.6.2

# OpenTelemetry instrumentation
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation>=0.42b0

# For OpenAI support (optional)
openai>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.0.0
mypy>=1.7.0